+++ src/1e00.c
@@ -4,7 +4,7 @@
 
 // cdd on wrong register. Maybe C++ compiler difference?
 #ifdef NON_MATCHING
-void func_1e00(unk_struct_2f9b0* arg0, long long arg1) {
+void func_1e00(unk_struct_2f9b0* arg0, u64 arg1) {
     // INCLUDE_ASM("asm/nonmatching", func_1e00);
     arg0->unk1B0 = arg1;
 }
